+++ UProfiler-Server/Services/VersionInfo.cs
@@ -27,6 +27,6 @@ public static class VersionInfo
             }
         }
 
-        return "1.1.0";
+        return "1.1.1";
     }
 }